refactor(scalar): explicit from_limbs/to_limbs via ScalarExt
Replace From/Into/RefInto<[u64; 4]> bounds with explicit limb
conversions and mod-order Montgomery round-trip.

Closes #228
