+++ crates/proof-of-sql/src/base/slice_ops/inner_product_test.rs
@@ -1,5 +1,5 @@
 use super::*;
-use crate::base::scalar::{test_scalar::TestScalar, ScalarExt};
+use crate::base::scalar::{test_scalar::TestScalar, Scalar};
 
 #[test]
 fn test_inner_product_with_bytes_basic() {
