+++ crates/proof-of-sql/src/sql/proof_exprs/inequality_expr_test.rs
@@ -7,7 +7,7 @@ use crate::{
         },
         math::decimal::Precision,
         posql_time::{PoSQLTimeUnit, PoSQLTimeZone},
-        scalar::{Scalar, ScalarExt},
+        scalar::Scalar,
     },
     proof_primitive::inner_product::curve_25519_scalar::Curve25519Scalar,
     sql::{
