+++ crates/proof-of-sql/src/sql/proof_exprs/test_utility.rs
@@ -137,7 +137,7 @@ pub fn const_varbinary(val: &[u8]) -> DynProofExpr {
 
 /// Create a constant scalar value. Used if we don't want to specify column types.
 pub fn const_scalar<S: Scalar, T: Into<S>>(val: T) -> DynProofExpr {
-    DynProofExpr::new_literal(LiteralValue::Scalar(val.into().into()))
+    DynProofExpr::new_literal(LiteralValue::Scalar(val.into().to_limbs()))
 }
 
 /// # Panics
